+++ lib/processor/compress_test.go
@@ -22,7 +22,9 @@ package processor
 
 import (
 	"bytes"
+	"compress/flate"
 	"compress/gzip"
+	"compress/zlib"
 	"os"
 	"reflect"
 	"testing"
@@ -64,6 +66,101 @@ func TestCompressGZIP(t *testing.T) {
 		var buf bytes.Buffer
 
 		zw := gzip.NewWriter(&buf)
+		zw.Write(input[i])
+		zw.Close()
+
+		exp = append(exp, buf.Bytes())
+	}
+
+	if reflect.DeepEqual(input, exp) {
+		t.Fatal("Input and exp output are the same")
+	}
+
+	proc, err := NewCompress(conf, nil, testLog, metrics.DudType{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	msgs, res := proc.ProcessMessage(types.NewMessage(input))
+	if len(msgs) != 1 {
+		t.Error("Compress failed")
+	} else if res != nil {
+		t.Errorf("Expected nil response: %v", res)
+	}
+	if act := msgs[0].GetAll(); !reflect.DeepEqual(exp, act) {
+		t.Errorf("Unexpected output: %s != %s", act, exp)
+	}
+}
+
+func TestCompressZLIB(t *testing.T) {
+	conf := NewConfig()
+	conf.Compress.Algorithm = "zlib"
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+
+	input := [][]byte{
+		[]byte("hello world first part"),
+		[]byte("hello world second part"),
+		[]byte("third part"),
+		[]byte("fourth"),
+		[]byte("5"),
+	}
+
+	exp := [][]byte{}
+
+	for i := range input {
+		var buf bytes.Buffer
+
+		zw := zlib.NewWriter(&buf)
+		zw.Write(input[i])
+		zw.Close()
+
+		exp = append(exp, buf.Bytes())
+	}
+
+	if reflect.DeepEqual(input, exp) {
+		t.Fatal("Input and exp output are the same")
+	}
+
+	proc, err := NewCompress(conf, nil, testLog, metrics.DudType{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	msgs, res := proc.ProcessMessage(types.NewMessage(input))
+	if len(msgs) != 1 {
+		t.Error("Compress failed")
+	} else if res != nil {
+		t.Errorf("Expected nil response: %v", res)
+	}
+	if act := msgs[0].GetAll(); !reflect.DeepEqual(exp, act) {
+		t.Errorf("Unexpected output: %s != %s", act, exp)
+	}
+}
+
+func TestCompressFlate(t *testing.T) {
+	conf := NewConfig()
+	conf.Compress.Algorithm = "flate"
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+
+	input := [][]byte{
+		[]byte("hello world first part"),
+		[]byte("hello world second part"),
+		[]byte("third part"),
+		[]byte("fourth"),
+		[]byte("5"),
+	}
+
+	exp := [][]byte{}
+
+	for i := range input {
+		var buf bytes.Buffer
+
+		zw, err := flate.NewWriter(&buf, conf.Compress.Level)
+		if err != nil {
+			t.Fatal(err)
+		}
 		zw.Write(input[i])
 		zw.Close()